requirements/base.txt
fastapi==0.104.1
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0.post1
gunicorn==21.2.0
asgi-ratelimit==0.10.0
structlog==23.2.0
